requirements.txt
langchain==0.0.352
google-generativeai==0.3.1
streamlit==1.29.0
python-dotenv==1.0.0
numpy==1.26.2
certifi>=2024.2.2
youtube-transcript-api==0.6.1
